ws/pyproject.toml
[tool.poetry]
name = "levity-ws"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "levity_ws"}]

[tool.poetry.dependencies]
python = "^3.10"
starlette = "^0.26.1"
python-dotenv = "^1.0.0"
aio-pika = "^9.0.5"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
pytest-asyncio = "^0.21.0"
pika = "^1.3.1"
httpx = "^0.24.0"
async-asgi-testclient = "^1.4.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
